ContactBookApi/requirements.txt
flask
psycopg2
python-dotenv
customtkinter
Pillow
requests
flask-smorest